+++ back/CampusForum/CampusForum/Models/Follow.cs
@@ -9,9 +9,9 @@ namespace CampusForum.Models
     public class Follow
     {
         [Key]
-        public int id { get; set; }
-        public int user_id { get; set; }
-        public int follower_id { get; set; }
+        public long id { get; set; }
+        public long user_id { get; set; }
+        public long follower_id { get; set; }
         public DateTime gmt_create { set; get; }
         public DateTime gmt_modified { set; get; }
